chore(global): убрать standalone сборку
Jira issue: VEGA-674
